+++ plugins/rust/examples/ramdisk.rs
@@ -52,6 +52,10 @@ impl Server for RamDisk {
     }
 
     fn open(_readonly: bool) -> Result<Box<dyn Server>> {
+        // The following line will print the message below when
+        // nbdkit is run with the -v flag:
+        //   nbdkit: ramdisk[1]: debug: connection opened: readonly=false
+        debug!("connection opened: readonly={}", _readonly);
         Ok(Box::<RamDisk>::default())
     }